resume-mentor/pyproject.toml
[project]
name = "resume-mentor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "groq>=1.0.0",
    "plotly>=6.5.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
